feat: Add API for task management
Introduce changes to the README.md file, updating the project description to emphasize its focus on task management. It also elaborates on the technology stack utilized in the project, including AWS Lambda, API Gateway, DocumentDB, and Cognito. Additionally, it provides instructions for project deployment, outlines the project structure, mentions other technologies and practices like TDD and DDD, and provides guidance on running tests and contributing to the project.